+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+- Add experimental session replay capturing on Xbox development kits ([#1378](https://github.com/getsentry/sentry-unreal/pull/1378))
 - Add SDK shutdown timeout as plugin setting for native desktop and console platforms ([#1379](https://github.com/getsentry/sentry-unreal/pull/1379))
 
 ### Fixes